requirements.tests.txt
pytest
pytest-cov
pytest-django
faker